+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 - Support to retrieve the height from a DNL marker segment.
 - Support to encode and decode mixed interleaved mode scans.
 - The unit tests are now based on Google test instead of MSTest and can be used on all platforms.
+- Additional checks when encoding/decoding color transformation. Only 3 components, 8/16 bits per sample, lossless and interleave mode line/sample are supported.
 
 ### Fixed